Added snapshot tests for BLIK component
